requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.0
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
discord.py==1.3.3
Django==3.0.6
ffprobe==0.5
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
image==1.5.32
isort==4.3.21
lazy-object-proxy==1.4.3
logzero==1.5.0
lyricwikia==0.1.11
mccabe==0.6.1
multidict==4.7.5
mutagen==1.44.0
numpy==1.18.4
pafy==0.5.5
pandas==1.0.4
pathlib==1.0.1
Pillow==7.1.2
playsound==1.2.2
psutil==5.7.0
pycparser==2.20
pylint==2.5.0
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
rendascii==1.0.0a2.post2
requests==2.23.0
six==1.14.0
soupsieve==2.0
spotdl==1.2.6
spotipy==2.12.0
sqlparse==0.3.1
titlecase==0.12.0
toml==0.10.0
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.25.9
websockets==8.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.8
yarl==1.4.2